Bump swift version and fix linter issues
