+++ .travis.yml
@@ -1,7 +1,7 @@
 sudo: required
 dist: trusty
 language: generic
-osx_image: xcode9.3
+osx_image: xcode10.1
 
 # Dependencies
 before_install: chmod +x ./.travis/deps.sh
