+++ CELLULAR.podspec
@@ -1,12 +1,13 @@
 Pod::Spec.new do |spec|
-    spec.version     = '4.1.0'
-    spec.module_name = 'CELLULAR'
-    spec.name        = 'CELLULAR'
-    spec.summary     = 'Collection of µ-frameworks and utility classes/extensions used in CELLULAR projects.'
-    spec.homepage    = 'https://www.cellular.de'
-    spec.authors     = { 'CELLULAR GmbH' => 'iosdev@cellular.de' }
-    spec.license     = { :type => 'MIT', :file => 'LICENSE' }
-    spec.source      = { :git => 'https://github.com/cellular/cellular-swift.git', :tag => spec.version.to_s }
+    spec.version       = '5.0.0'
+    spec.swift_version = '4.2'
+    spec.module_name   = 'CELLULAR'
+    spec.name          = 'CELLULAR'
+    spec.summary       = 'Collection of µ-frameworks and utility classes/extensions used in CELLULAR projects.'
+    spec.homepage      = 'https://www.cellular.de'
+    spec.authors       = { 'CELLULAR GmbH' => 'iosdev@cellular.de' }
+    spec.license       = { :type => 'MIT', :file => 'LICENSE' }
+    spec.source        = { :git => 'https://github.com/cellular/cellular-swift.git', :tag => spec.version.to_s }
 
     # Deployment Targets
     spec.ios.deployment_target     = '9.0'
