+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:4.1
+// swift-tools-version:4.2
 import PackageDescription
 
 let package = Package(
